+++ templates/typescript/server.ts
@@ -1,3 +1,4 @@
+import chalk from "chalk";
 import * as express from "express";
 import * as _ from "lodash";
 import { VoxaPlatform } from "voxa";
@@ -68,4 +69,11 @@ if (config.server.hostSkill) {
 expressApp.use(accountLinkingRoutes);
 {{/if}}
 
-expressApp.listen(config.server.port);
+expressApp.listen(config.server.port, () => {
+  console.log("");
+  console.log(chalk.green("Listening on:"));
+  console.log("");
+  Object.keys(routes).forEach((route) => {
+    console.log(chalk.green(`localhost:${config.server.port}${route}`));
+  });
+});