+++ test/dialogflow.spec.ts
@@ -1,9 +1,16 @@
+import fs = require("fs-extra");
 import { expect } from "chai";
 import * as _ from "lodash";
 import * as path from "path";
 import { configurations } from "./mocha.spec";
 
 configurations.forEach(interactionFile => {
+  if (
+    !_.includes(interactionFile.platforms, "dialogflow") &&
+    !interactionFile.dialogflowSpreadsheets
+  ) {
+    return;
+  }
   describe(`${interactionFile.name} Dialogflow`, () => {
     let agent: any;
 
@@ -12,21 +19,24 @@ configurations.forEach(interactionFile => {
         return this.skip();
       }
 
-      agent = await require(path.join(
-        __dirname,
+      const agentPath = path.join(
+        path.dirname(interactionFile.interactionFileName),
         interactionFile.speechPath,
         "dialogflow/production/agent.json"
-      ));
+      );
+      agent = await JSON.parse((await fs.readFile(agentPath)).toString());
     });
 
     describe("GOOGLE_ASSISTANT_WELCOME", () => {
       let intent: any;
       before(async () => {
-        intent = await require(path.join(
-          __dirname,
+        const intentPath = path.join(
+          path.dirname(interactionFile.interactionFileName),
           interactionFile.speechPath,
           "dialogflow/production/intents/GOOGLE_ASSISTANT_WELCOME.json"
-        ));
+        );
+
+        intent = JSON.parse((await fs.readFile(intentPath)).toString());
       });
 
       it("should generate a GOOGLE_ASSISTANT_WELCOME intent", () => {
@@ -46,16 +56,19 @@ configurations.forEach(interactionFile => {
       let intent: any;
       let intentUtterance: any;
       before(async () => {
-        intent = await require(path.join(
-          __dirname,
+        const intentPath = path.join(
+          path.dirname(interactionFile.interactionFileName),
           interactionFile.speechPath,
           "dialogflow/production/intents/NumberIntent.json"
-        ));
-        intentUtterance = await require(path.join(
-          __dirname,
+        );
+        const utterancesPath = path.join(
+          path.dirname(interactionFile.interactionFileName),
           interactionFile.speechPath,
           "dialogflow/production/intents/NumberIntent_usersays_en.json"
-        ));
+        );
+
+        intent = JSON.parse((await fs.readFile(intentPath)).toString());
+        intentUtterance = JSON.parse((await fs.readFile(utterancesPath)).toString());
       });
 
       it("should set slotRequired for the first slot to be false", () => {
@@ -84,11 +97,12 @@ configurations.forEach(interactionFile => {
     describe("JokeIntent", () => {
       let intent: any;
       before(async () => {
-        intent = await require(path.join(
-          __dirname,
+        const intentPath = path.join(
+          path.dirname(interactionFile.interactionFileName),
           interactionFile.speechPath,
           "dialogflow/production/intents/JokeIntent.json"
-        ));
+        );
+        intent = JSON.parse((await fs.readFile(intentPath)).toString());
       });
       it("should set webhookUsed to false", () => {
         expect(intent.webhookUsed).to.be.false;
@@ -104,11 +118,12 @@ configurations.forEach(interactionFile => {
       let intent: any;
 
       before(async () => {
-        intent = await require(path.join(
-          __dirname,
+        const intentPath = path.join(
+          path.dirname(interactionFile.interactionFileName),
           interactionFile.speechPath,
           "dialogflow/production/intents/DateIntent.json"
-        ));
+        );
+        intent = JSON.parse((await fs.readFile(intentPath)).toString());
       });
 
       it("should set slotRequired for the first slot to be true", () => {