+++ CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [2.2.1] - 2019-09-17
+
+### Added
+
+- Added prompt to check if the project should be generated in the same folder or not.
+
+### Fixed
+
+- When building the alexa interaction model there was a bug where even when empty the Dialog property was still being included, that's fixed now
 
 ## [2.2.0] - 2019-06-10